Move findPackage, getPackageFileName, and programError to /src/lib
